requirements.txt
baidu-aip==2.2.13.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
get==2019.3.22
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
post==2019.3.22
public==2019.3.22
pymongo==3.7.2
query-string==2019.3.22
request==2019.3.22
requests==2.21.0
urllib3==1.24.1
Werkzeug==0.15.2
